Create State.java
This class stores a board and its number of attacking queens (fitness). The class is used for the genetic algorithm search. When put into a priority queue, the states are sorted reversely by fitness. This allows States in the priority queue to be pulled and the ones left in are the best States in terms of fitness (lower fitness is better).